refactor: move botstats to stats command
