+++ src/commands/stats.ts
@@ -10,20 +10,30 @@ import {
     MessageEditOptions,
     MessageOptions,
 } from "discord.js";
+import { cpu } from "node-os-utils";
+import * as os from "os";
 import { addCooldown, getResponse, onCooldown } from "../utils/cooldownhandler";
 import prisma from "../utils/database/database";
 import redis from "../utils/database/redis";
 import { getStats } from "../utils/economy/utils";
+import { MStoTime } from "../utils/functions/date";
 import { getCommandUses } from "../utils/karma/utils";
 import { Categories, Command, NypsiCommandInteraction } from "../utils/models/Command";
 import { CustomEmbed } from "../utils/models/EmbedBuilders";
+// @ts-expect-error typescript doesnt like opening package.json
+import { version } from "../../package.json";
+import { workerCount } from "../events/message";
+import { aliasesSize, commandsSize } from "../utils/commandhandler";
+import { NypsiClient } from "../utils/models/Client";
+import { mentionQueue } from "../utils/users/utils";
 
 const cmd = new Command("stats", "view your nypsi stats", Categories.MONEY);
 
 cmd.slashEnabled = true;
 cmd.slashData
     .addSubcommand((economy) => economy.setName("economy").setDescription("view your economy stats"))
-    .addSubcommand((commands) => commands.setName("commands").setDescription("view your command usage stats"));
+    .addSubcommand((commands) => commands.setName("commands").setDescription("view your command usage stats"))
+    .addSubcommand((bot) => bot.setName("bot").setDescription("view nypsi's stats"));
 
 async function run(message: Message | (NypsiCommandInteraction & CommandInteraction), args: string[]) {
     const send = async (data: MessageOptions | InteractionReplyOptions) => {
@@ -249,6 +259,88 @@ async function run(message: Message | (NypsiCommandInteraction & CommandInteract
         return pageManager();
     };
 
+    const botStats = async () => {
+        const systemUptime = MStoTime(os.uptime() * 1000);
+        const uptime = MStoTime(message.client.uptime);
+        const totalMem = Math.round(os.totalmem() / 1024 / 1024);
+        const freeMem = Math.round(os.freemem() / 1024 / 1024);
+        const memUsage = Math.round(totalMem - freeMem);
+        const cpuUsage = await cpu.usage();
+
+        const usersDb = await prisma.user.count();
+        const economyDb = await prisma.economy.count();
+        const premDb = await prisma.premium.count();
+
+        const client = message.client as NypsiClient;
+
+        const clusterCount = client.cluster.count;
+        const currentCluster = client.cluster.id;
+        const currentShard = message.guild.shardId;
+
+        const userCount: number = await client.cluster
+            .broadcastEval("this.users.cache.size")
+            .then((res) => res.reduce((a, b) => a + b));
+        const guildCount: number = await client.cluster
+            .broadcastEval("this.guilds.cache.size")
+            .then((res) => res.reduce((a, b) => a + b));
+
+        let collections = 0;
+        let mentions = 0;
+
+        for (const mention of mentionQueue) {
+            if (mention.type == "collection") {
+                collections++;
+            } else if (mention.type == "mention") {
+                mentions++;
+            }
+        }
+
+        const embed = new CustomEmbed(message.member)
+            .setHeader(`nypsi stats | cluster: ${currentCluster + 1}/${clusterCount}`, client.user.avatarURL())
+            .addField(
+                "bot",
+                "**server count** " +
+                    guildCount.toLocaleString() +
+                    "\n" +
+                    "**users cached** " +
+                    userCount.toLocaleString() +
+                    "\n" +
+                    "**total commands** " +
+                    commandsSize +
+                    "\n" +
+                    "**total aliases** " +
+                    aliasesSize,
+                true
+            )
+            .addField(
+                "database",
+                `**users** ${usersDb.toLocaleString()}\n**economy** ${economyDb.toLocaleString()}\n**premium** ${premDb.toLocaleString()}`,
+                true
+            )
+            .addField(
+                "mention queue",
+                "**total** " +
+                    mentionQueue.length.toLocaleString() +
+                    "\n-- **collections** " +
+                    collections.toLocaleString() +
+                    "\n-- **mentions** " +
+                    mentions.toLocaleString() +
+                    "\n-- **workers** " +
+                    workerCount.toLocaleString(),
+                true
+            )
+            .addField(
+                "system",
+                `**memory** ${memUsage.toLocaleString()}mb/${totalMem.toLocaleString()}mb\n**cpu** ${cpuUsage}%\n**uptime** ${systemUptime}`,
+                true
+            )
+            .addField("cluster", `**uptime** ${uptime}`, true);
+
+        embed.setFooter({ text: `v${version} | shard: ${currentShard}` });
+
+        return send({ embeds: [embed] });
+    };
+
     if (args.length == 0) {
         return normalStats();
     } else if (args[0].toLowerCase() == "global" && message.author.id == "672793821850894347") {
@@ -390,6 +482,8 @@ async function run(message: Message | (NypsiCommandInteraction & CommandInteract
         return normalStats();
     } else if (args[0].toLowerCase().includes("command") || args[0].toLowerCase().includes("cmd")) {
         return commandStats();
+    } else if (args[0].toLowerCase().includes("bot") || args[0].toLowerCase().includes("nypsi")) {
+        return botStats();
     } else {
         return normalStats();
     }
